Use faker-provided email for testing
